requirements.txt
boto==2.46.1
click==6.7
Collectfast==0.5.2
colorthief==0.2.1
contextlib2==0.5.4
dj-database-url==0.4.2
Django==1.11
django-ckeditor==5.2.1
django-markdown-deux==1.0.5
django-redis-cache==1.7.1
django-rq==0.9.5
django-s3direct==0.4.9
django-storages==1.5.2
django-taggit==0.22.0
feedparser==5.2.1
gunicorn==19.7.1
html2text==2016.9.19
markdown2==2.3.3
olefile==0.44
Pillow==4.1.0
psycopg2==2.7.1
python-dateutil==2.6.0
python-oembed==0.2.4
pytz==2017.2
raven==6.0.0
redis==2.10.5
requests==2.13.0
rq==0.7.1
six==1.10.0
sorl-thumbnail==12.3
